requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
flask
gitpython
gpytranslate
hachoir
heroku3
httpx==0.27.2
motor
numpy
opencv-python
pillow==9.5.0
psutil
py-tgcalls==0.9.7
pykeyboard
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
uvloop
yt-dlp
youtube-search
youtube-search-python
pyromod
wget
blackpink
